Migrated common functions to generalFunctions.py
